refactor: queue heartbeat evaluation on redis
