+++ .env.example
@@ -55,6 +55,7 @@ REDIS_CLIENT=phpredis
 REDIS_HOST=127.0.0.1
 REDIS_PASSWORD=null
 REDIS_PORT=6379
+HEARTBEAT_QUEUE=heartbeat
 
 MAIL_MAILER=log
 MAIL_SCHEME=null
