+++ composer.json
@@ -59,7 +59,7 @@
         ],
         "dev": [
             "Composer\\Config::disableProcessTimeout",
-            "bunx concurrently -c \"#93c5fd,#c4b5fd,#fb7185,#fdba74\" \"php artisan serve\" \"php artisan queue:listen --tries=1\" \"php artisan pail --timeout=0\" \"bun run dev\" --names=server,queue,logs,vite"
+            "bunx concurrently -c \"#93c5fd,#c4b5fd,#fb7185,#fdba74,#facc15\" \"php artisan serve\" \"php artisan queue:listen --tries=1\" \"php artisan queue:listen redis --queue=heartbeat --tries=1\" \"php artisan pail --timeout=0\" \"bun run dev\" --names=server,queue,heartbeat,logs,vite"
         ],
         "test": [
             "@php artisan test"
